Rename wanda -> checks
